Unit Testing concurrent software is not straight forward at all. Running every test in two threads means that the assertions may be outdated by the time they are executed if the second thread changes the state. Will need to discuss this with Kostas. In the mean time, JUnit tests are being written for the Mutator class.
Another issue to solve is running the unit tests on the mutated files since the mutated file will have a descriptive name but the class name will be the same. This may involve moving the actual files once they have passed their unit tests and integrating the mutated files directly into the project. This shouldn't be too strenuous if we are testing the files manually, but could be quite tricky if the unit tests are to be run automatically.