After the mutation, it would be nice if it could refactor the java file to have a separate name, save it to the project and then perform identical unit tests on it. At the moment, the current plan would be to do this manually which isn’t too much hassle if you have separate mutation versions for each tested class which gets overwritten by a new mutation.
Created some single-thread unit tests for the whole banking system in JUnit tests. Figured out what the hell I'm meant to be doing by looking at the bankSystem ScenarioTests. I have targeted the concurrent methods of deposit, withdraw and transfer for each of the relevant classes. They have been saved as bankSystemUnitTests.

Added another method to the rtxc operator - Thread.interrupt() and Thread.await(time) - which are used in various of the banking classes and can now be tested with the mutator for the first set of results. Had to update the regex to include the option of multiple parameters, as this was necessary for the await() method.